rewrite Container, CoverImage and HeroPost component
