+++ components/container.tsx
@@ -1,7 +1,10 @@
 import React from 'react'
+import { Box } from 'rebass'
 
-const Container: React.FC = ({ children }) => {
-  return <div className="container mx-auto px-5">{children}</div>
+export const Container: React.FC = ({ children }) => {
+  return (
+    <Box mx="auto" maxWidth={['1280px']} px={1} sx={{ width: '100%' }}>
+      {children}
+    </Box>
+  )
 }
-
-export default Container
