+++ components/footer.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import Container from './container'
+import { Container } from './Container'
 
 const Footer: React.FC = () => {
   return (
