+++ components/post-preview.tsx
@@ -5,7 +5,7 @@ import Author from '../types/author'
 
 import Avatar from './atoms/Avatar'
 import DateFormatter from './date-formatter'
-import CoverImage from './cover-image'
+import { CoverImage } from './CoverImage'
 
 type Props = {
   title: string
