+++ pages/blog.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import Head from 'next/head'
 
-import Container from '../components/container'
+import { Container } from '../components/Container'
 import MoreStories from '../components/more-stories'
-import HeroPost from '../components/hero-post'
+import { HeroPost } from '../components/HeroPost'
 import Layout from '../components/layout'
 import { getAllPosts } from '../lib/api'
 import { CMS_NAME } from '../lib/constants'
@@ -28,7 +28,6 @@ const Index = ({ allPosts }: Props) => {
               title={heroPost.title}
               coverImage={heroPost.coverImage}
               date={heroPost.date}
-              author={heroPost.author}
               slug={heroPost.slug}
               excerpt={heroPost.excerpt}
             />
@@ -44,6 +43,7 @@ export default Index
 
 // eslint-disable-next-line @typescript-eslint/require-await
 export const getStaticProps = async () => {
+  // TODO author prop is unnessesary
   const allPosts = getAllPosts([
     'title',
     'date',
