+++ pages/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import Head from 'next/head'
 
-import Container from '../components/container'
+import { Container } from '../components/Container'
 import Layout from '../components/layout'
 import { SITE_NAME } from '../lib/constants'
 import Hero from '../components/organisms/Hero'
